Update Manga Converter.py
The old code was taking a while to generate the .pdf at the end, in the case of over 4000 pages it could take up to 2 hours, which isn't practical.
This new code will compress each image as it goes (currently 35% quality, you can amend this) and builds a pdf for each chapter which is far faster. It will continue to combine the .pdfs as it goes so you're left with one .pdf that contains all chapters and compressed images. I've stretched the images out to full a4 too, this is so it looks best when read on a kindle.